Safari-609.1.11.1/LayoutTests/svg/animations/animate-currentColor-expected.txt

SVG 1.1 dynamic animation tests

Tests animation on 'currentColor'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 128
PASS colorComponents[2] is 64
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 128
PASS colorComponents[3] is 0
PASS successfullyParsed is true

TEST COMPLETE
